feat(dde-control-center): 删除无用代码，修改代码中出现的警告,泄露，空指针问题
删除一些无用代码，以及一些由于变量为初始化，赋值类型不一致导致的警告,泄露，空指针问题问题

Log: 控制中心废弃代码删除,代码出现警告,泄露，空指针问题

Bug: https://pms.uniontech.com/zentao/task-view-27932.html

+++ src/frame/modules/display/monitorsground.cpp
@@ -36,8 +36,8 @@ using namespace dcc::display;
 
 const int MARGIN_W = 20;
 const int MARGIN_H = 10;
-const double VIEW_WIDTH = 400.;
-const double VIEW_HEIGHT = 200.;
+const int VIEW_WIDTH = 400;
+const int VIEW_HEIGHT = 200;
 
 MonitorsGround::MonitorsGround(QWidget *parent)
     : QFrame(parent),
@@ -124,11 +124,11 @@ void MonitorsGround::monitorMoved(MonitorProxyWidget *pw)
     qDebug() << Q_FUNC_INFO << pw->name();
 
     const double scale = screenScale();
-    const int offsetX = VIEW_WIDTH / 2 - (m_viewPortWidth * scale) / 2 + MARGIN_W;
-    const int offsetY = VIEW_HEIGHT / 2 - (m_viewPortHeight * scale) / 2 + MARGIN_H;
+    const double offsetX = VIEW_WIDTH / 2 - (m_viewPortWidth * scale) / 2 + MARGIN_W;
+    const double offsetY = VIEW_HEIGHT / 2 - (m_viewPortHeight * scale) / 2 + MARGIN_H;
 
-    pw->setMovedX((pw->pos().x() - offsetX) / scale);
-    pw->setMovedY((pw->pos().y() - offsetY) / scale);
+    pw->setMovedX(static_cast<int>((pw->pos().x() - offsetX) / scale));
+    pw->setMovedY(static_cast<int>((pw->pos().y() - offsetY) / scale));
 
     // ensure screens is 贴合但不相交
     ensureWidgetPerfect(pw);
@@ -160,23 +160,22 @@ void MonitorsGround::adjust(MonitorProxyWidget *pw)
 
     const double scale = screenScale();
 
-    const int offsetX = VIEW_WIDTH / 2 - (m_viewPortWidth * scale) / 2 + MARGIN_W;
-    const int offsetY = VIEW_HEIGHT / 2 - (m_viewPortHeight * scale) / 2 + MARGIN_H;
+    const double offsetX = VIEW_WIDTH / 2 - (m_viewPortWidth * scale) / 2 + MARGIN_W;
+    const double offsetY = VIEW_HEIGHT / 2 - (m_viewPortHeight * scale) / 2 + MARGIN_H;
 
-    const int w = scale * pw->w();
-    const int h = scale * pw->h();
-    const int x = scale * pw->x();
-    const int y = scale * pw->y();
+    const double w = scale * pw->w();
+    const double h = scale * pw->h();
+    const double x = scale * pw->x();
+    const double y = scale * pw->y();
 
     if(bSingle)
     {
-        pw->setGeometry((width()-w)/2, (height()-h)/2, scale * pw->w(), scale * pw->h());
+        pw->setGeometry(static_cast<int>((width()-w)/2), static_cast<int>((height()-h)/2), static_cast<int>(scale * pw->w()), static_cast<int>(scale * pw->h()));
         this->setEnabled(false);//单屏时不允许鼠标拖动 不然以前的机制会导致窗体重算引发方大
     }
     else
         pw->setGeometry(x + offsetX, y + offsetY, w, h);
     pw->update();
-
     //解决设置1.25缩放，自定义拖动dp和edp缩略图有残影问题
     Q_EMIT requestUpdateWidget();
 }
